Move EnvironmentError to quasar.fs
