+++ interface/src/main/scala/quasar/main/logging.scala
@@ -19,10 +19,9 @@ package quasar.main
 import slamdata.Predef.{PartialFunction, Some, Unit}
 
 import quasar.Planner, Planner.PlannerError
-import quasar.connector.EnvironmentError
 import quasar.effect.{Capture, Failure}
 import quasar.fp.free
-import quasar.fs.{FileSystemError, PhysicalError, UnhandledFSError}
+import quasar.fs.{EnvironmentError, FileSystemError, PhysicalError, UnhandledFSError}
 import quasar.fs.mount.MountingError
 import quasar.fs.mount.module.Module
