+++ web/src/main/scala/quasar/api/ToApiError.scala
@@ -19,7 +19,6 @@ package quasar.api
 import slamdata.Predef._
 import quasar.{Data, DataCodec, Planner, SemanticError}
 import quasar.RenderTree.ops._
-import quasar.connector.EnvironmentError
 import quasar.fp._
 import quasar.fp.ski._
 import quasar.fs._
